requirements.txt
semantic_kernel
python-dotenv
serpapi